Add ramstkfunction table mock
